requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
autogen==0.7.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
fast-depends==2.4.12
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.63.0
packaging==24.2
pyautogen==0.7.4
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
PyMuPDF==1.25.3
python-docx==1.1.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
termcolor==2.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
